Rename lua_registry to lua_ecs
